crawler/requirements.txt
playwright==1.52.0
beautifulsoup4==4.13.4
requests==2.32.3
supabase==2.15.2
python-dotenv==1.1.0
lxml==5.4.0
anthropic==0.55.0
